pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "transform-movie-to-text"
version = "0.1.0"
description = "A UI app to extract text from video files using Whisper."
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
dependencies = [
    "openai-whisper",
    "google-genai",
    "flet[all]>=0.21.0",
    "flet-desktop",
    "psutil",
    "torch",
    "torchvision",
    "torchaudio",
    "ollama",
    "openai",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-mock",
    "ruff",
]

[tool.ruff]
line-length = 150
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"] # Basic errors, formatting, and imports
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.uv]
index = [
  { name = "pytorch", url = "https://download.pytorch.org/whl/cu121", explicit = true }
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
torchaudio = { index = "pytorch" }
